migrate views
