+++ package.json
@@ -4,9 +4,10 @@
   "private": true,
   "homepage": "./",
   "dependencies": {
-    "@figurl/core-components": "https://github.com/magland/figurl-core-components/releases/download/0.1.3/figurl-core-components-v0.1.2.tgz",
-    "@figurl/core-utils": "magland/figurl-core-utils.git#0.1.0",
-    "@figurl/interface": "magland/figurl-interface.git#0.1.1",
+    "@figurl/core-components": "https://github.com/magland/figurl-core-components/releases/download/0.1.4/figurl-core-components-v0.1.4.tgz",
+    "@figurl/core-utils": "https://github.com/magland/figurl-core-utils/releases/download/0.1.1/figurl-core-utils-v0.1.1.tgz",
+    "@figurl/interface": "https://github.com/magland/figurl-interface/releases/download/0.1.3/figurl-interface-v0.1.3.tgz",
+    "@figurl/spike-sorting-views": "https://github.com/magland/figurl-spike-sorting-views/releases/download/0.1.1/figurl-spike-sorting-views-v0.1.1.tgz",
     "@fortawesome/fontawesome-svg-core": "^6.1.1",
     "@fortawesome/free-regular-svg-icons": "^6.1.2",
     "@fortawesome/free-solid-svg-icons": "^6.1.1",
@@ -28,7 +29,7 @@
     "crypto": "npm:crypto-browserify",
     "deck.gl": "^8.8.4",
     "github-markdown-css": "^5.1.0",
-    "mathjs": "^10.6.4",
+    "mathjs": "^11.2.1",
     "object-hash": "^3.0.0",
     "path": "npm:path-browserify",
     "react": "^18.2.0",
