+++ src/View.tsx
@@ -1,15 +1,15 @@
 import { TiledImageComponent } from 'libraries/component-tiled-image';
 import { AnnotationsView } from 'libraries/view-annotations';
-import { AutocorrelogramsView } from 'libraries/view-autocorrelograms';
+import { AutocorrelogramsView } from '@figurl/spike-sorting-views';
 import { AverageWaveformsView } from 'libraries/view-average-waveforms';
 import { CompositeView } from 'libraries/view-composite';
 import { ConfusionMatrixView } from 'libraries/view-confusion-matrix';
 import { ConsoleView } from 'libraries/view-console';
-import { CrossCorrelogramsView } from 'libraries/view-cross-correlograms';
+import { CrossCorrelogramsView } from '@figurl/spike-sorting-views';
 import { ElectrodeGeometryView } from 'libraries/view-electrode-geometry';
 import { EpochsView } from 'libraries/view-epochs';
 import { ExperimentalSelector1View } from 'libraries/view-experimental-selector-1';
-import { LiveCrossCorrelogramsView } from 'libraries/view-live-cross-correlograms';
+import { LiveCrossCorrelogramsView } from '@figurl/spike-sorting-views';
 import { LiveEvaluateFunctionView } from 'libraries/view-live-evaluate-function';
 import { LiveTracesView } from 'libraries/view-live-traces';
 import { MainLayoutView } from 'libraries/view-main-layout';
