--- src/libraries/context-sorting-curation/index.ts
@@ -1,4 +1,0 @@
-export type { default as SortingCurationAction } from './SortingCurationAction'
-export { default as SortingCurationContext, sortingCurationReducer, useSortingCuration } from './SortingCurationContext'
-export type { SortingCuration } from './SortingCurationContext'
-
