--- src/libraries/context-unit-selection/UnitSelectionTypes.ts
@@ -1,6 +1,0 @@
-export type SortingRule = {
-    columnName: string,
-    sortAscending: boolean
-}
-
-export type SortingCallback = (rules: SortingRule[]) => (number | string)[]
