--- src/libraries/context-unit-selection/index.ts
@@ -1,7 +1,0 @@
-export { idToNum } from './idToNum'
-export { sortIds } from './sortIds'
-export { COPY_STATE, default as UnitSelectionContext, defaultUnitSelection, DESELECT_ALL, INITIALIZE_UNITS, TOGGLE_SELECT_ALL, UNIQUE_SELECT_FIRST, UNIQUE_SELECT_LAST, UNIQUE_SELECT_NEXT, UNIQUE_SELECT_PREVIOUS, unitSelectionReducer, UPDATE_SORT_FIELDS, useSelectedUnitIds } from './UnitSelectionContext'
-export type { UnitSelectionAction, UnitSelectionState } from './UnitSelectionContext'
-export { allUnitSelectionState, getCheckboxClickHandlerGenerator, voidClickHandler } from './UnitSelectionFunctions'
-export type { SortingRule } from './UnitSelectionTypes'
-
