--- src/libraries/view-autocorrelograms/index.ts
@@ -1,4 +1,0 @@
-export { default as AutocorrelogramsView } from './AutocorrelogramsView'
-export { isAutocorrelogramsViewData } from './AutocorrelogramsViewData'
-export type { AutocorrelogramsViewData } from './AutocorrelogramsViewData'
-export { default as CorrelogramPlot, determineTickLocationsMsec } from './CorrelogramPlot'
