--- src/libraries/view-cross-correlograms/index.ts
@@ -1,4 +1,0 @@
-export { default as CrossCorrelogramsView, determinePlotSizeForSquareMatrixGrid } from './CrossCorrelogramsView'
-export { isCrossCorrelogramsViewData } from './CrossCorrelogramsViewData'
-export type { CrossCorrelogramsViewData } from './CrossCorrelogramsViewData'
-
