--- src/libraries/view-live-cross-correlograms/index.ts
@@ -1,3 +1,0 @@
-export { default as LiveCrossCorrelogramsView } from './LiveCrossCorrelogramsView'
-export { isLiveCrossCorrelogramsViewData } from './LiveCrossCorrelogramsViewData'
-export type { LiveCrossCorrelogramsViewData } from './LiveCrossCorrelogramsViewData'
