+++ src/libraries/view-unit-locations/UnitLocationsWidget.tsx
@@ -2,7 +2,7 @@ import { useSelectedElectrodes } from 'libraries/context-recording-selection'
 import { BaseCanvas } from '@figurl/core-components'
 import { pointInRect, RectangularRegion, rectangularRegionsIntersect, transformPoint, Vec2, Vec4 } from '@figurl/core-components'
 import { colorForUnitId } from '@figurl/core-utils'
-import { idToNum, useSelectedUnitIds } from 'libraries/context-unit-selection'
+import { idToNum, useSelectedUnitIds } from '@figurl/spike-sorting-views'
 import { useCallback, useMemo } from 'react'
 import { AffineTransform, applyAffineTransform, applyAffineTransformInv, detAffineTransform } from 'libraries/view-unit-similarity-matrix'
 import { useWheelZoom } from 'libraries/view-unit-similarity-matrix'
