+++ src/libraries/view-unit-metrics-graph/UnitMetricsGraphViewChild.tsx
@@ -1,12 +1,12 @@
 import { useUnitMetricSelection } from 'libraries/context-unit-metrics-selection';
 import { PGPlot, PlotGrid } from '@figurl/core-components';
 import { Splitter } from '@figurl/core-components';
-import { idToNum, INITIALIZE_UNITS, sortIds, useSelectedUnitIds } from 'libraries/context-unit-selection';
+import { idToNum, INITIALIZE_UNITS, sortIds, useSelectedUnitIds } from '@figurl/spike-sorting-views';
 import { VerticalScrollView } from '@figurl/core-components';
 import { ToolbarItem, ViewToolbar } from 'libraries/ViewToolbar';
 import { FunctionComponent, useEffect, useMemo, useState } from 'react';
 import { FaMinus, FaPlus } from 'react-icons/fa';
-import { determinePlotSizeForSquareMatrixGrid } from 'libraries/view-cross-correlograms';
+import { determinePlotSizeForSquareMatrixGrid } from '@figurl/spike-sorting-views';
 import UnitMetricPlot, { UnitMetricPlotProps } from './UnitMetricPlot';
 import { UMGMetric, UnitMetricsGraphViewData } from './UnitMetricsGraphViewData';
